+++ planning/services/configurationProvider.js
@@ -3,12 +3,22 @@ angular.module('90Tech.planning')
         var self = this;
 
         self.BASE_SIZE = 10;
+        self.MAX_PARALLEL = 3;
         self.strings = {
             nothing_to_show: 'Rien à afficher'
         };
 
+        self.parallelText = 'parallel items';
+
+        this.setParallelText = function(text){
+            self.parallelText = ''+text;
+        }
+
         this.setBaseSize = function(size){
             self.BASE_SIZE = size;
+        };
+        this.setMaxParallel = function(number){
+            self.MAX_PARALLEL = number;
         };
 
         this.setString= function(key, value){
@@ -16,7 +26,7 @@ angular.module('90Tech.planning')
         };
 
         this.$get = [function(){
-            return {strings: self.strings, BASE_SIZE : self.BASE_SIZE};
+            return {strings: self.strings, BASE_SIZE : self.BASE_SIZE, MAX_PARALLEL : self.MAX_PARALLEL, parallelText : self.parallelText};
         }];
 
     });