+++ planning/templates.js
@@ -2,174 +2,97 @@ angular.module('90Tech.planning').run(['$templateCache', function($templateCache
   'use strict';
 
   $templateCache.put('planning/directives/planning-left-column/planning-left-column.html',
-    "<div class=\"left-column\" ng-switch=\"planningLeftColumn.mode\">\r" +
-    "\n" +
-    "    <div class=\"planning-head\"></div>\r" +
-    "\n" +
-    "    <div class=\"left-body\">\r" +
-    "\n" +
-    "    <div class=\"planning-2pc\"></div>\r" +
-    "\n" +
-    "        <div class=\"days-body\" ng-switch-when=\"week\">\r" +
-    "\n" +
-    "            <div class=\"dayName row8 b-b animate\"\r" +
-    "\n" +
-    "                 ng-repeat=\"day in planningLeftColumn.days\"\r" +
-    "\n" +
-    "                    ng-class=\"{today: planning.isToday(day.dayOfYear())}\"\r" +
-    "\n" +
-    "                    ng-click=\"planning.dayCallback({$day:day})\">\r" +
-    "\n" +
-    "                <h4>\r" +
-    "\n" +
-    "                    {{day | format:'dddd' | capitalize}}<br>\r" +
-    "\n" +
-    "                    <small>{{day | format:'ll'}}</small>\r" +
-    "\n" +
-    "                </h4>\r" +
-    "\n" +
-    "            </div>\r" +
-    "\n" +
-    "        </div>\r" +
-    "\n" +
-    "        <div class=\"days-body\" ng-switch-when=\"day\">\r" +
-    "\n" +
-    "            <div class=\"dayName row8 b-b animate\"\r" +
-    "\n" +
-    "                 ng-repeat=\"col in planningLeftColumn.column\">\r" +
-    "\n" +
-    "                <h4>\r" +
-    "\n" +
-    "                    {{col}}\r" +
-    "\n" +
-    "                </h4>\r" +
-    "\n" +
-    "            </div>\r" +
-    "\n" +
-    "        </div>\r" +
-    "\n" +
-    "    </div>\r" +
-    "\n" +
-    "    <div style=\"height:10px\"><!-- compensate for scrollbar --></div>\r" +
-    "\n" +
+    "<div class=\"left-column\" ng-switch=\"planningLeftColumn.mode\">\n" +
+    "    <div class=\"planning-head\"></div>\n" +
+    "    <div class=\"left-body\">\n" +
+    "    <div class=\"planning-2pc\"></div>\n" +
+    "        <div class=\"days-body\" ng-switch-when=\"week\">\n" +
+    "            <div class=\"dayName row8 b-b animate\"\n" +
+    "                 ng-repeat=\"day in planningLeftColumn.days\"\n" +
+    "                    ng-class=\"{today: planning.isToday(day.dayOfYear())}\"\n" +
+    "                    ng-click=\"planning.dayCallback({$day:day})\">\n" +
+    "                <h4>\n" +
+    "                    {{day | format:'dddd' | capitalize}}<br>\n" +
+    "                    <small>{{day | format:'ll'}}</small>\n" +
+    "                </h4>\n" +
+    "            </div>\n" +
+    "        </div>\n" +
+    "        <div class=\"days-body\" ng-switch-when=\"day\">\n" +
+    "            <div class=\"dayName row8 b-b animate\"\n" +
+    "                 ng-repeat=\"col in planningLeftColumn.column\">\n" +
+    "                <h4>\n" +
+    "                    {{col}}\n" +
+    "                </h4>\n" +
+    "            </div>\n" +
+    "        </div>\n" +
+    "    </div>\n" +
+    "    <div style=\"height:10px\"><!-- compensate for scrollbar --></div>\n" +
     "</div>"
   );
 
 
   $templateCache.put('planning/directives/planning-line/planning-line.html',
-    "<div class=\"b-b b-r all-day day-hour\"\r" +
+    "<div class=\"b-b b-r all-day day-hour\"\n" +
+    "     ng-style=\"{width: line.calcWidth(line.zoom)}\"\n" +
+    "     ng-repeat=\"n in [] | range:line.range\"\n" +
+    "     ng-dblclick=\"line.clickEvent(n, $event)\">\n" +
+    "    <span class=\"half-hour\"></span>\n" +
     "\n" +
-    "     ng-style=\"{width: line.calcWidth(line.zoom)}\"\r" +
+    "</div>\n" +
     "\n" +
-    "     ng-repeat=\"n in [] | range:line.range\"\r" +
+    "<div class=\"event\"\n" +
+    "     ng-repeat=\"event in line._events\"\n" +
+    "     ng-style=\"event.style\"\n" +
+    "     ng-class=\"{'continued-before': event.continuedBefore, 'continued-after': event.continuedAfter}\"\n" +
+    "     ng-click=\"planning.eventCallback({'event':event})\"\n" +
+    "     tooltip-append-to-body=\"true\"\n" +
+    "     uib-tooltip=\"{{event.tooltip}}\">\n" +
+    "    <!--<span class=\"calendar-urgency bg-pink\"></span>-->\n" +
     "\n" +
-    "     ng-dblclick=\"line.clickEvent(n, $event)\">\r" +
+    "    <div class=\"event-line-container\" style=\"\">\n" +
+    "        <div class=\"event-line\" ng-style=\"{'background-color': event.color}\" ng-if=\"!event.continuedBefore\"></div>\n" +
+    "    </div>\n" +
     "\n" +
-    "    <span class=\"half-hour\"></span>\r" +
-    "\n" +
-    "\r" +
-    "\n" +
-    "</div>\r" +
-    "\n" +
-    "\r" +
-    "\n" +
-    "<div class=\"event\"\r" +
-    "\n" +
-    "     ng-repeat=\"event in line._events\"\r" +
-    "\n" +
-    "     ng-style=\"event.style\"\r" +
-    "\n" +
-    "     ng-class=\"{'continued-before': event.continuedBefore, 'continued-after': event.continuedAfter}\"\r" +
-    "\n" +
-    "     ng-click=\"planning.eventCallback({'event':event})\"\r" +
-    "\n" +
-    "     tooltip-append-to-body=\"true\"\r" +
-    "\n" +
-    "     uib-tooltip=\"{{event.tooltip}}\">\r" +
-    "\n" +
-    "    <!--<span class=\"calendar-urgency bg-pink\"></span>-->\r" +
-    "\n" +
-    "\r" +
-    "\n" +
-    "    <div class=\"event-line-container\" style=\"\">\r" +
-    "\n" +
-    "        <div class=\"event-line\" ng-style=\"{'background-color': event.color}\" ng-if=\"!event.continuedBefore\"></div>\r" +
-    "\n" +
-    "    </div>\r" +
-    "\n" +
-    "\r" +
-    "\n" +
-    "    <div class=\"title-container\"><span>{{event.title}}</span></div>\r" +
-    "\n" +
+    "    <div class=\"title-container\"><span>{{event.title}}</span></div>\n" +
     "</div>"
   );
 
 
   $templateCache.put('planning/directives/planning-top-row/planning-top-row.html',
-    "<div class=\"planning-top-row\">\r" +
-    "\n" +
-    "    <div ng-repeat=\"hour in planningTopRow.hours\" class=\"day-hour\"\r" +
-    "\n" +
-    "         ng-style=\"{width: planningTopRow.calcWidth(planningTopRow.zoom), margin: planningTopRow.calcMargin(planningTopRow.zoom, $index)}\">\r" +
-    "\n" +
-    "        <h4>{{hour | format:'HH:mm'}}</h4>\r" +
-    "\n" +
-    "    </div>\r" +
-    "\n" +
-    "</div>\r" +
-    "\n" +
-    "\r" +
+    "<div class=\"planning-top-row\">\n" +
+    "    <div ng-repeat=\"hour in planningTopRow.hours\" class=\"day-hour\"\n" +
+    "         ng-style=\"{width: planningTopRow.calcWidth(planningTopRow.zoom), margin: planningTopRow.calcMargin(planningTopRow.zoom, $index)}\">\n" +
+    "        <h4>{{hour | format:'HH:mm'}}</h4>\n" +
+    "    </div>\n" +
+    "</div>\n" +
     "\n"
   );
 
 
   $templateCache.put('planning/directives/planning/planning.html',
-    "<zl-planning-left-column mode=\"planning.mode\" position=\"planning.position\" day-field=\"planning.dayField\"\r" +
-    "\n" +
-    "                         events=\"planning.sortedEvents\"></zl-planning-left-column>\r" +
-    "\n" +
-    "<div class=\"\" style=\"height:100%\">\r" +
-    "\n" +
-    "    <div zl-horizontal-scroll scroll-left=\"planning.currentTimeToPixels()\" style=\"height:100%\">\r" +
-    "\n" +
-    "        <div ng-style=\"{width: planning.width}\" class=\"planning-body\">\r" +
-    "\n" +
-    "            <zl-planning-top-row mode=\"planning.mode\" zoom=\"planning.zoom\" position=\"planning.position\"\r" +
-    "\n" +
-    "                                 day-start=\"planning._dayStart\" day-end=\"planning._dayEnd\"></zl-planning-top-row>\r" +
-    "\n" +
-    "            <div class=\"hour-cursor\" ng-style=\"{left: planning.currentTimeToPixels()+'px'}\"\r" +
-    "\n" +
-    "                 ng-if=\"planning.isCurrent() && planning.isInDayRange()\">\r" +
-    "\n" +
-    "                <div class=\"hour-caret\"></div>\r" +
-    "\n" +
-    "            </div>\r" +
-    "\n" +
-    "            <div class=\"planning-2pc\"></div>\r" +
-    "\n" +
-    "            <div ng-if=\"!planning.events.length\" style=\"position:fixed;left:50%;margin:auto\">{{'nothing_to_show' |\r" +
-    "\n" +
-    "                strPlanning}}\r" +
-    "\n" +
-    "            </div>\r" +
-    "\n" +
-    "            <zl-planning-line\r" +
-    "\n" +
-    "                    zoom=\"planning.zoom\"\r" +
-    "\n" +
-    "                    day-start=\"planning._dayStart\" day-end=\"planning._dayEnd\"\r" +
-    "\n" +
-    "                    ng-repeat=\"i in planning.keys(planning.sortedEvents)\" class=\"day b-b\"\r" +
-    "\n" +
-    "                    ng-class=\"{today: planning.isToday(i)}\" events=\"planning.getEvents(i)\"\r" +
-    "\n" +
-    "                    click-callback=\"planning.clickCallbackWrapper($hour, $minutes, i)\"></zl-planning-line>\r" +
-    "\n" +
-    "        </div>\r" +
-    "\n" +
-    "    </div>\r" +
-    "\n" +
+    "<zl-planning-left-column mode=\"planning.mode\" position=\"planning.position\" day-field=\"planning.dayField\"\n" +
+    "                         events=\"planning.sortedEvents\"></zl-planning-left-column>\n" +
+    "<div class=\"\" style=\"height:100%\">\n" +
+    "    <div zl-horizontal-scroll scroll-left=\"planning.currentTimeToPixels()\" style=\"height:100%\">\n" +
+    "        <div ng-style=\"{width: planning.width}\" class=\"planning-body\">\n" +
+    "            <zl-planning-top-row mode=\"planning.mode\" zoom=\"planning.zoom\" position=\"planning.position\"\n" +
+    "                                 day-start=\"planning._dayStart\" day-end=\"planning._dayEnd\"></zl-planning-top-row>\n" +
+    "            <div class=\"hour-cursor\" ng-style=\"{left: planning.currentTimeToPixels()+'px'}\"\n" +
+    "                 ng-if=\"planning.isCurrent() && planning.isInDayRange()\">\n" +
+    "                <div class=\"hour-caret\"></div>\n" +
+    "            </div>\n" +
+    "            <div class=\"planning-2pc\"></div>\n" +
+    "            <div ng-if=\"!planning.events.length\" style=\"position:fixed;left:50%;margin:auto\">{{'nothing_to_show' |\n" +
+    "                strPlanning}}\n" +
+    "            </div>\n" +
+    "            <zl-planning-line\n" +
+    "                    zoom=\"planning.zoom\"\n" +
+    "                    day-start=\"planning._dayStart\" day-end=\"planning._dayEnd\"\n" +
+    "                    ng-repeat=\"i in planning.keys(planning.sortedEvents)\" class=\"day b-b\"\n" +
+    "                    ng-class=\"{today: planning.isToday(i)}\" events=\"planning.getEvents(i)\"\n" +
+    "                    click-callback=\"planning.clickCallbackWrapper($hour, $minutes, i)\"></zl-planning-line>\n" +
+    "        </div>\n" +
+    "    </div>\n" +
     "</div>"
   );