+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 Version numbers correspond to `bower.json` version
 
+# 1.0.12
+
+## Features
+
+- Add a maximum parallel items
+- Add empty space in each case
+- Add sourcemap
+
 # 1.0.0
 
 ## Features